+++ internal/storage/storageutil/client_helper_test.go
@@ -17,30 +17,10 @@ package storageutil
 import (
 	"net/url"
 	"testing"
-	"time"
 
-	mountpkg "github.com/googlecloudplatform/gcsfuse/internal/mount"
 	"github.com/jacobsa/oglematchers"
 	. "github.com/jacobsa/ogletest"
 )
-
-// GetDefaultStorageClientConfig is only for test, making the default endpoint
-// non-nil, so that we can create dummy tokenSource while unit test.
-func GetDefaultStorageClientConfig() (clientConfig StorageClientConfig) {
-	return StorageClientConfig{
-		ClientProtocol:      mountpkg.HTTP1,
-		MaxConnsPerHost:     10,
-		MaxIdleConnsPerHost: 100,
-		HttpClientTimeout:   800 * time.Millisecond,
-		MaxRetryDuration:    30 * time.Second,
-		RetryMultiplier:     2,
-		UserAgent:           "gcsfuse/unknown (Go version go1.20-pre3 cl/474093167 +a813be86df) (GCP:gcsfuse)",
-		Endpoint:            &url.URL{},
-		KeyFile:             DummyKeyFile,
-		TokenUrl:            "",
-		ReuseTokenFromUrl:   true,
-	}
-}
 
 func TestClientHelper(t *testing.T) { RunTests(t) }
